moved timesync from p2p dir
